+++ CesiumGeometry/include/CesiumGeometry/AxisAlignedBox.h
@@ -1,8 +1,11 @@
 #pragma once
 
-#include <CesiumGeometry/Library.h>
+#include "Library.h"
 
+#include <glm/common.hpp>
 #include <glm/vec3.hpp>
+
+#include <vector>
 
 namespace CesiumGeometry {
 
@@ -119,6 +122,30 @@ struct CESIUMGEOMETRY_API AxisAlignedBox final {
     return position.x >= this->minimumX && position.x <= this->maximumX &&
            position.y >= this->minimumY && position.y <= this->maximumY &&
            position.z >= this->minimumZ && position.z <= this->maximumZ;
+  }
+
+  /**
+   * @brief Creates a tight-fitting, axis-aligned bounding box that contains all
+   * of the input positions.
+   *
+   * @param position The positions.
+   * @returns An axis-aligned bounding box containing the positions.
+   */
+  static AxisAlignedBox
+  fromPositions(const std::vector<glm::dvec3>& positions) {
+    if (positions.size() == 0) {
+      return AxisAlignedBox();
+    }
+
+    glm::dvec3 min = positions[0];
+    glm::dvec3 max = positions[0];
+
+    for (size_t i = 1; i < positions.size(); i++) {
+      min = glm::min(min, positions[i]);
+      max = glm::max(max, positions[i]);
+    }
+
+    return AxisAlignedBox(min.x, min.y, min.z, max.x, max.y, max.z);
   }
 };